Solver promise not really generic so move out of utils.
